line/pyproject.toml
[project]
name = "embsys-line-classroom"
version = "1.0.0"
description = "LINE Beacon attendance backend for Smart Classroom"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "psycopg[binary]>=3.1",
]

[project.scripts]
serve = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
